# cedar-monitor-server-application/src/main/resources/assets/swagger-api/swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "Better Metadata - Better Science\n\nThe monitor server reports on the rest of CEDAR: how deep the work queues are, how much each\nstore holds, whether a given server considers itself healthy, and what artifact a given string\nrefers to.\n\nEverything here is read-only and needs the monitor read permission. The counts are meant to be\nread against each other: the workspace graph is authoritative, and the search index disagreeing\nwith it is how a half-finished reindex shows itself.\n",
    "title" : "CEDAR Monitor Server API",
    "version" : "2.9.10-SNAPSHOT"
  },
  "paths" : {
...
    "/logs/boards" : {
      "get" : {
        "description" : "The catalogue of predefined questions, each one a saved query spec a client can load into the same controls and edit. The catalogue is served from here so a client's list cannot drift from what the query engine supports. Specs carry no time range: the caller supplies one, and `defaultRangeMinutes` is only a suggested starting window. A few boards join the two log tables and so cannot be a spec; those name a route in `endpoint`, which the caller reads instead of posting `spec`. Requires the monitor read permission.",
        "operationId" : "boards",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/Board"
                  }
                }
              }
            },
            "description" : "The board catalogue"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "List the predefined Insight boards",
        "tags" : [ "Log query" ]
      }
    },
    "/logs/coverage" : {
      "get" : {
        "description" : "Per log table: the columns a query may name, which of them can be grouped and which aggregated, and the rows and time window the table actually holds. Some columns were added after logging began and so carry values only for recent rows; the notes on a column say so. A client reads this to state its own caveats rather than showing a column that looks broken. Requires the monitor read permission.",
        "operationId" : "coverage",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoverageResult"
                }
              }
            },
            "description" : "The queryable surface and the data behind it"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The log database could not be read"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Describe what is queryable and what is actually present",
        "tags" : [ "Log query" ]
      }
    },
    "/logs/db-share" : {
      "get" : {
        "description" : "One row per handler and component over the range, slowest in total first, with the request count, the summed handler time, the summed Cypher time and the share the second is of the first. The answer comes back in the same shape as a structured query, so a client renders it with the same table. It needs a route of its own only because it joins the request log to the Cypher log, which a spec cannot express. Requires the monitor read permission.",
        "operationId" : "dbShare",
        "parameters" : [ {
          "description" : "Inclusive lower bound, as an ISO-8601 instant in UTC. Defaults to seven days before the upper bound.",
          "in" : "query",
          "name" : "from",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Exclusive upper bound, as an ISO-8601 instant in UTC. Defaults to now.",
          "in" : "query",
          "name" : "to",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "How many rows to return. Defaults to 100 and is capped at 500.",
          "in" : "query",
          "name" : "limit",
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QueryResult"
                }
              }
            },
            "description" : "One row per handler, slowest in total first"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LogQueryError"
                }
              }
            },
            "description" : "A bound was not an ISO-8601 instant, or `from` was not before `to`"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The log database could not be read"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Compare each handler's total time against the database time underneath it",
        "tags" : [ "Log query" ]
      }
    },
...
    "/logs/facets/{column}" : {
      "get" : {
        "description" : "What a filter dropdown needs: every value the column took over the range, most frequent first, with how often each occurred. Only columns the engine marks facetable can be asked for. The result says whether the value list was capped. Requires the monitor read permission.",
        "operationId" : "facet",
        "parameters" : [ {
          "description" : "The column to enumerate, named as the coverage route names it.",
          "in" : "path",
          "name" : "column",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Which log table to read. Defaults to the request log.",
          "in" : "query",
          "name" : "table",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Inclusive lower bound, as an ISO-8601 instant in UTC. Defaults to 24 hours before the upper bound.",
          "in" : "query",
          "name" : "from",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Exclusive upper bound, as an ISO-8601 instant in UTC. Defaults to now.",
          "in" : "query",
          "name" : "to",
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/FacetResult"
                }
              }
            },
            "description" : "The column's distinct values and their counts"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LogQueryError"
                }
              }
            },
            "description" : "The table or column is not known or not facetable, a bound was not an ISO-8601 instant, or `from` was not before `to`"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The log database could not be read"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "List the distinct values of one column, with their counts",
        "tags" : [ "Log query" ]
      }
    },
    "/logs/query" : {
      "post" : {
        "description" : "The one query surface the Explorer, the pivot view and every Insight board are built on. The body is a structured spec rather than SQL text: column names are resolved against an allowlist and values are bound, so a spec naming a column the engine does not know is rejected rather than run. An empty `groupBy` returns raw rows newest first, paged by the cursor the previous result handed back; a non-empty one returns an aggregate. The result carries its own provenance, so a caller can tell an exact answer from a histogram-approximate one and a complete result from a capped one. It is POST rather than GET because a spec with several filters and a metric list outgrows a URL. Requires the monitor read permission.",
        "operationId" : "query",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LogQuerySpec"
              }
            }
          },
          "description" : "The query to run. Every field is optional and is normalized before use: `table` defaults to the request log, `source` selects the raw log tables or the hourly rollups, and `limit` is capped by the engine.",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/QueryResult"
                }
              }
            },
            "description" : "The columns, the rows and the result's provenance"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/LogQueryError"
                }
              }
            },
            "description" : "The spec named an unknown table, column, metric or operator, or its cursor could not be read"
          },
          "401" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "Unauthorized"
          },
          "403" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The caller lacks the monitor read permission"
          },
          "500" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CedarError"
                }
              }
            },
            "description" : "The log database could not be read"
          }
        },
        "security" : [ {
          "api_key" : [ ]
        } ],
        "summary" : "Run a structured query over the log tables",
        "tags" : [ "Log query" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Board" : {
        "type" : "object",
        "properties" : {
          "defaultRangeMinutes" : {
            "type" : "integer",
            "format" : "int32"
          },
          "endpoint" : {
            "type" : "string"
          },
          "group" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          },
          "question" : {
            "type" : "string"
          },
          "spec" : {
            "$ref" : "#/components/schemas/LogQuerySpec"
          },
          "title" : {
            "type" : "string"
          }
        }
      },
...
      "CoverageResult" : {
        "type" : "object",
        "properties" : {
          "notes" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "tables" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/TableCoverage"
            }
          }
        }
      },
...
      "LogQueryError" : {
        "type" : "object",
        "additionalProperties" : false,
        "description" : "The body the log query routes answer a bad request with. It is not the CEDAR error envelope\nevery other route in this service uses: `badRequest` in LogQueryResource builds this one\ndirectly. Documented as it behaves; converging it on the shared envelope would change what\na caller receives.\n",
        "properties" : {
          "error" : {
            "type" : "string",
            "description" : "What was wrong with the request, in the words the query layer raised."
          }
        },
        "required" : [ "error" ]
      },
...
    },
    "securitySchemes" : {
      "api_key" : {
        "description" : "Your CEDAR API key, using the format 'apiKey <yourApiKey>'",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  }
}